refactor: each fetch command scores its own sources
fetch:metars owns metar rows, fetch:vatsim owns vatsim/event/logon_estimate
(and reuses its own vatsim-data.json download for VATSIM_POPULAR),
fetch:bookings owns booking rows. calc:scores is gone.
