+++ app/Console/Commands/UpdateData.php
@@ -40,9 +40,6 @@ class UpdateData extends Command
         $this->info('>> fetch:vatsim running');
         $this->call('fetch:vatsim');
 
-        $this->info('>> calc:scores running');
-        $this->call('calc:scores');
-
         $this->info('> Done with all commands in ' . round(microtime(true) - $processTime) . ' seconds!');
 
     }
